requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
python-telegram-bot==20.7
psycopg[binary]==3.2.13
python-dotenv==1.0.0
requests==2.31.0
